feat: multi-session catch-up — SINAD rework, GPS/SFI/solar services, nets, propagation scoring
SINAD DSP rework (ale2g_modem.cpp):
  Symbol decision on full 64-sample block unchanged; SINAD now computed
  on centered 32-sample window [16..47] using time-domain total power vs
  Goertzel signal power — eliminates ~25% off-fundamental energy leak at
  sub-symbol edges. Clean loopback: 30 dB (was ~6-7). Verified by new
  test_sinad_rework.cpp. goertzel_power gains optional window-length param.

GPS / SFI / solar-position services (new):
  gps_service.{h,cpp}, sfi_service.{h,cpp}, solar_position.{h,cpp} —
  time-of-day and propagation-condition awareness for LQA channel ranking.
  freq_select_manager.{h,cpp} — frequency selection manager.
  lqa_exchange.{h,cpp} — LQA exchange protocol support.
  Tests: test_gps_parsers, test_sfi_service, test_solar_position,
  test_propagation_scoring.

Nets / channel files:
  nets/ directory with 8 HFLINK per-net .ale files
  (USA/EUR/UK/JA/AUS/ITU1/ITU2/ITU3).

Protocol / state machine / controller:
  Broad updates across ale_state_machine.cpp, ale_controller.cpp,
  ale_bridge.cpp, ale_data_store.cpp — multiple sessions of fixes and
  feature additions per commit history context.

GUI:
  app.js, index.html, styles.css — UI updates tracking bridge changes.

Build / config:
  CMakeLists.txt, ale.conf — build system and station config updates.
  .gitignore: exclude root-level *.raw/*.wav test-capture artifacts.

Docs:
  docs/SINAD_REWORK_HANDOFF.md, Features/Feature_Measurement_LQA.md,
  CONTEXT.md, ARCHITECTURAL ANALYSIS.md, README.md.

+++ tests/link/integration/test_bilateral_lqa.cpp
@@ -312,7 +312,7 @@ static void test_full_handshake_with_bilateral_lqa()
     // ── Step 6: Feed SAM's ACK to JOE ───────────────────────────────────────
     // Additional SAM TX words (ACK) were appended to sam_log after response received.
     // Drive JOE time forward to cover the ACK window.
-    const uint32_t t_sam_ack_start = t_joe_first + joe_log.words.size() * Trw;
+    const uint32_t t_sam_ack_start = t_joe_first + static_cast<uint32_t>(joe_log.words.size()) * Trw;
     size_t joe_rx_idx = sam_log.words.size();  // JOE already got sam_log[0..n-1]
     uint32_t t = t_sam_ack_start;
     for (; joe_rx_idx < sam_log.words.size(); ++joe_rx_idx, t += Trw) {
